publish += location - kamangir/bolt#698

+++ .abcli/ferfereh.sh
@@ -4,32 +4,41 @@ function ferfereh() {
     local task=$(abcli_unpack_keyword $1 help)
 
     if [ $task == "help" ] ; then
-        abcli_show_usage "ferfereh publish" \
+        abcli_show_usage "ferfereh publish 3d-files|locations" \
             "publish ferfereh."
         return
     fi
 
     if [ "$task" == "publish" ] ; then
-        local gens="gen1 gen5 gen6"
-        local folders="3d gcode sketchup"
+        local what=$(abcli_clarify_input $2 3d-files)
 
-        abcli_log "-ferfereh: publish: $gens X $folders"
+        if [ "$what" == "3d-files" ] ; then
+            local gens="gen1 gen5 gen6"
+            local folders="3d gcode sketchup"
 
-        local folder
-        local gen
-        for folder in $folders ; do
-            abcli_log "-ferfereh: publish: $gens -> $folder"
+            abcli_log "-ferfereh: publish: $what: $gens X $folders"
 
-            rm -rfv $abcli_path_git/ferfereh/$folder
-            mkdir -p $abcli_path_git/ferfereh/$folder
+            local folder
+            local gen
+            for folder in $folders ; do
+                abcli_log "-ferfereh: publish: $gens -> $folder"
 
-            for gen in $gens ; do
-                cp -v $abcli_path_git/dart/ferfereh/$folder/$gen* $abcli_path_git/ferfereh/$folder
+                rm -rfv $abcli_path_git/ferfereh/$folder
+                mkdir -p $abcli_path_git/ferfereh/$folder
+
+                for gen in $gens ; do
+                    cp -v $abcli_path_git/dart/ferfereh/$folder/$gen* $abcli_path_git/ferfereh/$folder
+                done
             done
-        done
 
-        cd $abcli_path_git/ferfereh
-        git status
+            cd $abcli_path_git/ferfereh
+            git status
+        elif [ "$what" == "locations" ] ; then
+            echo "wip"
+        else
+            abcli_log_error "-ferfereh: publish: $what: not found."
+            return 1
+        fi
 
         return
     fi
